Revert "refactor log service (#292)"
This reverts commit e44d465b30e2b14c61f941929c6c3ce5b8c1c252.

+++ logservice/logpuller/log_puller_multi_span.go
@@ -63,7 +63,7 @@ func NewLogPullerMultiSpan(
 			if ts := pullerWrapper.tryUpdateGlobalResolvedTs(entry, subID); ts != 0 {
 				return consume(ctx, &common.RawKVEntry{
 					OpType: common.OpTypeResolved,
-					CRTs:   ts,
+					CRTs:   uint64(ts),
 				})
 			}
 			return nil
